Changed control to holding the spacebar to rotate
Also adjusted the algorithm for calculating numSpins. If the player rotates 360 degrees in air, numSpins increases by 1. If the player almost completes a circle when hitting the trampoline, that also counts as a spin. Everytime the player hits the trampoline, the rotation degree counter is reset to 0.

Also adjusted the codes inside "if(movex >/< 0) in response to the new control